+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ShowSessionsAction.java
@@ -46,6 +46,7 @@ public class ShowSessionsAction extends JiraWebActionSupport {
     public List<SessionEntity> getOpenSessions() {
         return scrumPokerSessionService.recent().stream()
             .filter(session -> session.getConfirmedVote() == null)
+            .filter(session -> !session.isCancelled())
             .filter(session -> getIssue(session.getIssueKey()) != null)
             .filter(session -> currentUserIsAllowedToSeeIssue(getIssue(session.getIssueKey())))
             .map(session -> sessionEntityTransformer.build(session, currentUser().getKey()))
@@ -54,7 +55,7 @@ public class ShowSessionsAction extends JiraWebActionSupport {
 
     public List<SessionEntity> getClosedSessions() {
         return scrumPokerSessionService.recent().stream()
-            .filter(session -> session.getConfirmedVote() != null)
+            .filter(session -> session.getConfirmedVote() != null || session.isCancelled())
             .filter(session -> getIssue(session.getIssueKey()) != null)
             .filter(session -> currentUserIsAllowedToSeeIssue(getIssue(session.getIssueKey())))
             .map(session -> sessionEntityTransformer.build(session, currentUser().getKey()))